Moved lcio dependent source in lcio directory
